Bytes (#23)
Introduce a `bytes` package that mirrors go's `bytes` package. Currently
supported functions are `Split` and `Repeat`.

Furthermore, a function `View` in `byteslice` is added to map `[]byte ->
seq[byte]`.

For `Split`, we require the separator to be non-empty. The Go
implementation handles this case by returning a string split at every
utf8 codepoint boundary: `Split("abc☺", "") -> "a", "b", "c", "☺"`
Do we want to mimic this behavior?

---------

Co-authored-by: João Pereira <joaopereira.19@gmail.com>